requirements.txt
pytest
fitparse
pydantic
PyYAML
